% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodiversity.R
\name{biodiversity}
\alias{biodiversity}
\title{Biodiversity indexes}
\usage{
biodiversity(database, occasional.species, species.cover.coefficient)
}
\arguments{
\item{database}{database with Frequency of occurrence (FO) and occasional species as 999. Rows are species and columns are surveys. Database class must be \emph{data.frame}}

\item{occasional.species}{Logical. TRUE if you want to take into account occasional species. If occasional species are considered, SRA correspond to "SRA_SC.fs.occ" (see \link[iPastoralist]{vegetation_abundance})}

\item{species.cover.coefficient}{To set only when "occasional.species=TRUE":coeffient that multiplies FO so that the number of total touches refer to 100. SRA values for Shannon diversity index derive from the "SRA_SC.fo.occ" (see \link[iPastoralist]{vegetation_abundance})}
}
\value{
biodiversity indexes per each survey
}
\description{
From a database with Frequency of occurrence (FO) and occasional species as 999 you can compute the following biodiversity indexes:
\itemize{
\item Species richness
\item Shannon diversity index (log2)
\item Effective number of species (ENS)
\item Shannon max
\item Equitability
}
}
\references{
Magurran, A.E., 1988. Ecological diversity and its measurement. Princeton University Press.
}
