% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster_group_aggregation.R
\name{clustGroupAggregation}
\alias{clustGroupAggregation}
\title{Computation of mean values for cluster groups}
\usage{
clustGroupAggregation(database)
}
\arguments{
\item{database}{}
}
\value{
database with cover-abundance values
}
\description{
\cr
}
\examples{
 Structure of the database to use as input:

      Rows: Vegetation survey
      Columns: species
      Headings and row names must be imported

        Group	  Spe1	  Spe2	...
          1     12.3	   0.0
          1     0.0     5.0
          1     2.0     0.0
          1     10.4    0.0
          2     0.0     20.7
          2     0.0     2.0
          2     2.3     0.0
          3     0.0     23.3
          3     14.3    0.0

  Note 1: the column of Group (e.g. Vegetation type, Facies..) 
           has to be in the first position
}
