% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_survey_order.R
\name{clustOrder}
\alias{clustOrder}
\title{Extraction of the position and/or the group of appartenance of survey in a dendrogram}
\usage{
clustOrder(database.cluster, cluster.hclust, cluster.group, cluster.number)
}
\arguments{
\item{database.cluster:}{database used for generate the cluster. Rows are survey and Columns are species}

\item{cluster.hclust:}{output from 'hclust' function of 'stats' package}

\item{cluster.group:}{Logical. TRUE if you want to specify the number of groups of the dendrogram.}

\item{cluster.number:}{To be specified only if 'cluster.group' is TRUE. Set the number of groups the dendrogram will be divided}
}
\value{
when 'cluster.group' = TRUE: a dataframe in which for each survey is specified the group id of appartenance and its position in the dendrogram
when 'cluster.group' = FALSE: a dataframe in which for each survey its position in the dendrogram
}
\description{
\cr
}
