% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecological_indexes.R
\name{ecological_indexes}
\alias{ecological_indexes}
\title{Computation of Ecological indicators (Landolt, Ellenberg)}
\usage{
ecological_indexes(
  database.vegetation,
  database.indexes,
  occasional.species,
  species.cover.coefficient,
  weight
)
}
\arguments{
\item{database.vegetation}{database with Frequency of occurrence (FO) and occasional species as 999. Rows are species and columns are surveys. The column of species names must be imported}

\item{database.indexes}{database with Ecological indicators, without the column of species names. NA values must indicated as 999}

\item{occasional.species}{Logical. TRUE if you want to take into account occasional species. If occasional species are considered, SRA correspond to "SRA_SC.fo.occ" (see \link[iPastoralism]{vegetation_abundance})}

\item{species.cover.coefficient}{only if "occasional.species=TRUE". Coeffient that multiplies FS so that the number of total touches refer to 100}

\item{weight}{Logical. TRUE if you want to weight Ecological indicators with SRA.}
}
\value{
database with Ecological indicators for each survey.
}
\description{
Computation of Ecological indicators, which can be either weighted or not weighted with SRA. Occasional species can be taken into
account. If occasional species are considered, SRA correspond to "SRA_SC.fo.occ" (see \link[iPastoralism]{vegetation_abundance})
}
\examples{
Structure of the database with ISQ to use as input:

      Headings must be imported. Species name column has not to be considered.

            N_Landolt T_Landolt
            2         5
            5         1
            4         999

      Note: For species without an Index value specify a value of 999.
}
