% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phytosociological_cover_transformation.R
\docType{data}
\name{method}
\alias{method}
\title{Transformation of phytosociological cover (\%) to cover-abundance values}
\format{
An object of class \code{character} of length 1.
}
\usage{
method
}
\arguments{
\item{database}{database with visually estimated cover of plant species. Rare species are added as '+". Rows are species and columns are surveys. The first column of the database reports the species names (Important: no space among words must exist). Database class must be \emph{data.frame}}

\item{method}{\itemize{
\item \strong{"BrBl"}:
\item \strong{"VanDerMaarel"}:
\item \strong{"TassTapp"}: modified Braun-Blanquet frequency/abundance scale
}}
}
\value{
database cover-abundance values
}
\description{
Species cover visually estimated in a phytosociological survey are transformed to:\cr
\itemize{
\item Braun-Blanquet cover-abundance scale \strong{(BrBl)} :
\item Van Der Maarel cover-abundance scale \strong{(VanDerMaarel)} :
\item Modified Braun-Blanquet frequency/abundance scale according to Tasser and Tappeiner (2005)**: \cr
}
}
\examples{
 Structure of the database to use as input:

      Rows: species
      Columns: vegetation survey
      Headings and row names must be imported

       	  Species   ril1     ril2
            spe1      +
            spe2              3
            spe3      2      34
            spe4      10
            spe5             20

          Note 1: For species without cover leave the cells empty (NA values)
          Note 2: rare species as '+'
}
\references{
\itemize{
\item Braun-Blanquet J (1932) Plant sociology. The study of plant communi- ties. GD Fuller and HS Conard (Eds.). Authorized English translations of 'Pflanzensoziologie'. 1st ed. Printed in the United States of America. New York and London: McGraw-Hill Book Co. Inc.\cr
\item Westhoff V, van der Maarel E (1973) The Braun-Blanquet approach. In: Whittaker RH (Ed.) Handbook of vegetation science, part 5, Classi- fication and ordination of communities. W Junk, The Hague: 617– 726. https://doi.org/10.1007/978-94-010-2701-4_20\cr
\item Tasser E, Tappeiner U (2005) New model to predict rooting in diverse plant community compositions. Ecological Modelling 185:195-211.\cr
}
}
\keyword{datasets}
