% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phytosociological_cover_transformation.R
\name{PhytoCover}
\alias{PhytoCover}
\title{Transformation of phytosociological cover (\%) to cover-abundance values}
\usage{
PhytoCover(database, method)
}
\arguments{
\item{database}{database with visually estimated cover of plant species. Rare species are added as '+". Rows are species and columns are surveys. The first column of the database reports the species names (Important: no space among words must exist). Database class must be \emph{data.frame}}

\item{method}{\itemize{
\item \strong{"BrBl"}:
\item \strong{"Maarel"}:
\item \strong{"TassTapp"}: modified Braun-Blanquet frequency/abundance scale
}}
}
\value{
database cover-abundance values
}
\description{
Species cover visually estimated in a phytosociological survey are transformed to:\cr
\itemize{
\item Braun-Blanquet cover-abundance scale \strong{(BrBl)} :
\item Van Der Maarel cover-abundance scale \strong{(Maarel)} :
\item Modified Braun-Blanquet frequency/abundance scale according to Tasser and Tappeiner (2005)**: \cr
}
}
\examples{
 Structure of the database to use as input:

      Rows: species
      Columns: vegetation survey
      Headings and row names must be imported

       	  Species   ril1     ril2
            spe1      +
            spe2              3
            spe3      2      34
            spe4      10
            spe5             20

          Note 1: For species without cover leave the cells empty (NA values)
          Note 2: rare species as '+'
}
\references{
\itemize{
\item Braun-Blanquet J (1932) Plant sociology. The study of plant communi- ties. GD Fuller and HS Conard (Eds.). Authorized English translations of 'Pflanzensoziologie'. 1st ed. Printed in the United States of America. New York and London: McGraw-Hill Book Co. Inc.\cr
\item Van der Maarel, E. (1979). Transformation of cover-abundance values in phytosociology and its effects on community similarity. Vegetatio, 39(2), 97-114.\cr
\item Tasser E, Tappeiner U (2005) New model to predict rooting in diverse plant community compositions. Ecological Modelling 185:195-211.\cr
}
}
