% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_ten_species.R
\name{first_ten_species}
\alias{first_ten_species}
\title{First ten species for dendrogram}
\usage{
first_ten_species(data_SRA_SC, export)
}
\arguments{
\item{data_SRA_SC}{database with abundance data expressed as SRA or SC, calculated with \link[iPastoralism]{vegetation_abundance}}

\item{export}{Logical. TRUE if you want export data in a csv format}
}
\description{
First ten species for dendrogram
}
