% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pastoral_value.R
\name{pastoral_value}
\alias{pastoral_value}
\title{Computation of Pastoral Value (PV)}
\usage{
pastoral_value(SRA_data, isq_data)
}
\arguments{
\item{SRA_data}{database with SRA values, either 'SRA_fs' or 'SRA_SC.fo.occ' (see \link[iPastoralist]{vegetation_abundance}). Database class must be \emph{data.frame}}

\item{isq_data}{database with ISQ values. Database class must be \emph{data.frame}}
}
\value{
database with PV
}
\description{
Computation of PV from a database with SRA, either SRA_fs or SRA_SC.fo.occ (see \link[iPastoralist]{vegetation_abundance})
}
